- feat ✨: change new create room, new update room and new delete room
- change new delete room route
- create building domain

+++ api/server/controllers/building.controller.js
@@ -1,100 +1,77 @@
 const { Response } = require('../../utils/response.util');
 const { CustomError, HandlerError } = require('../../utils/error.util');
 const { SUCCESS_STATUS } = require('../../constants/http-status.constant');
-const { NO_CONTENT_CODE, UNAUTHORIZED_CODE } = require('../../constants/http-status-code.constant');
+const { NO_CONTENT_CODE, UNAUTHORIZED_CODE, OK_CODE } = require('../../constants/http-status-code.constant');
 const { SOMETHING_WENT_WRONG, INVALID_REFRESH_TOKEN } = require('../../constants/error-message.constant');
-const { rooms, zones, buildings, waterZones } = require('../repositories/models');
+const { rooms, zones, buildings, waterZones, accommodations } = require('../repositories/models');
+const { CreateRoomScheme, CreateRoomDTO, UpdateRoomScheme, UpdateRoomDTO } = require('../domains/building.domain');
 const TokenList = require('./auth.controller');
-
 const building = async (req, res) => {
 	const getRefreshTokenFromHeader = await req.headers['x-refresh-token'];
 	try {
 		if (getRefreshTokenFromHeader && getRefreshTokenFromHeader in TokenList.TokenList) {
-			const roomsDetail = await rooms.findAll({
+			const room = await rooms.findAll({
 				include: [
 					{
 						model: zones,
-						attributes: ['id', 'name'],
 					},
 					{
 						model: waterZones,
-						attributes: ['id', 'name'],
 					},
 					{
 						model: buildings,
-						attributes: ['id', 'name'],
 					},
 				],
-				attributes: ['id', 'building_id', 'roomNo', 'roomType', 'waterNo', 'waterMeterNo', 'status'],
+				attributes: [
+					'id',
+					'building_id',
+					'roomNo',
+					'roomType',
+					'waterNo',
+					'waterMeterNo',
+					'electricityNo',
+					'electricityMeterNo',
+					'status',
+				],
 			});
-			if (!roomsDetail) {
+			if (!room) {
 				return HandlerError(res, CustomError(SOMETHING_WENT_WRONG));
 			} else {
-				res.status(200).json({ status: 'success', status_code: 200, result: roomsDetail });
+				return Response(res, SUCCESS_STATUS, OK_CODE, room);
 			}
 		} else {
 			return Response(res, INVALID_REFRESH_TOKEN, UNAUTHORIZED_CODE);
 		}
-	} catch (error) {
-		console.log(error);
+	} catch (err) {
+		return HandlerError(res, err);
 	}
 };
 
 const createRoom = async (req, res) => {
 	const getRefreshTokenFromHeader = await req.headers['x-refresh-token'];
+	const room = await CreateRoomScheme.validateAsync(req.body);
 	try {
-		const {
-			zoneId,
-			waterZoneId,
-			buildingId,
-			roomNo,
-			electricityNo,
-			electricityMeterNo,
-			waterNo,
-			waterMeterNo,
-			roomType,
-			status,
-		} = req.body;
-		const data = {
-			zoneId,
-			waterZoneId,
-			buildingId,
-			roomNo,
-			electricityNo,
-			electricityMeterNo,
-			waterNo,
-			waterMeterNo,
-			roomType,
-			status,
-		};
 		const findZone = await zones.findOne({
 			where: {
-				id: zoneId,
+				id: room.zoneId,
 			},
 		});
 		const findWaterZone = await waterZones.findOne({
 			where: {
-				id: waterZoneId,
+				id: room.waterZoneId,
 			},
 		});
 		const findbuilding = await buildings.findOne({
 			where: {
-				id: buildingId,
+				id: room.buildingId,
 			},
 		});
-		if (!findZone && findWaterZone && findbuilding) {
-			return HandlerError(res, CustomError(SOMETHING_WENT_WRONG));
-		}
-		if (findZone && !findWaterZone && findbuilding) {
-			return HandlerError(res, CustomError(SOMETHING_WENT_WRONG));
-		}
-		if (findZone && findWaterZone && !findbuilding) {
+		if (!findZone || !findWaterZone || !findbuilding) {
 			return HandlerError(res, CustomError(SOMETHING_WENT_WRONG));
 		}
 		if (findZone && findWaterZone && findbuilding) {
-			console.log(getRefreshTokenFromHeader in TokenList.TokenList);
 			if (getRefreshTokenFromHeader && getRefreshTokenFromHeader in TokenList.TokenList) {
-				const newRoom = await rooms.create(data);
+				const newRoom = await rooms.create(CreateRoomDTO(room));
 				if (newRoom) {
 					return Response(res, SUCCESS_STATUS, NO_CONTENT_CODE);
 				} else {
@@ -112,16 +89,21 @@ const createRoom = async (req, res) => {
 };
 
 const deleteRoom = async (req, res) => {
-	const id = await req.params.id;
+	const id = await req.body.id;
 	const getRefreshTokenFromHeader = await req.headers['x-refresh-token'];
 	try {
 		if (getRefreshTokenFromHeader && getRefreshTokenFromHeader in TokenList.TokenList) {
-			const room = await rooms.findOne({
-				where: {
-					id: id,
-				},
-			});
-			if (room) {
+			const room = await rooms.findAll({ where: { id: id } });
+			const accommodation = await accommodations.findAll({ where: { room_id: id } });
+			if (accommodation && room) {
+				await accommodations.update({ roomId: null }, { where: { room_id: id } });
+				await accommodations.update({ deleted: true }, { where: { room_id: id } });
+				await accommodations.update({ host: false }, { where: { room_id: id } });
+				rooms.destroy({ where: { id: id } });
+				console.log(accommodation);
+				return Response(res, SUCCESS_STATUS, NO_CONTENT_CODE);
+			}
+			if (room && !accommodation) {
 				rooms.destroy({ where: { id: id } });
 				return Response(res, SUCCESS_STATUS, NO_CONTENT_CODE);
 			}
@@ -136,18 +118,7 @@ const deleteRoom = async (req, res) => {
 const updateRoom = async (req, res) => {
 	const getRefreshTokenFromHeader = await req.headers['x-refresh-token'];
 	const id = await req.params.id;
-	const {
-		zoneId,
-		waterZoneId,
-		buildingId,
-		roomNo,
-		electricityNo,
-		electricityMeterNo,
-		waterNo,
-		waterMeterNo,
-		roomType,
-		status,
-	} = await req.body;
+	const data = await UpdateRoomScheme.validateAsync(req.body);
 	try {
 		const room = await rooms.findOne({
 			while: {
@@ -156,16 +127,19 @@ const updateRoom = async (req, res) => {
 		});
 		if (room) {
 			if (getRefreshTokenFromHeader && getRefreshTokenFromHeader in TokenList.TokenList) {
-				await rooms.update({ zoneId: zoneId }, { where: { id: id } });
-				await rooms.update({ waterZoneId: waterZoneId }, { where: { id: id } });
-				await rooms.update({ buildingId: buildingId }, { where: { id: id } });
-				await rooms.update({ roomNo: roomNo }, { where: { id: id } });
-				await rooms.update({ electricityNo: electricityNo }, { where: { id: id } });
-				await rooms.update({ electricityMeterNo: electricityMeterNo }, { where: { id: id } });
-				await rooms.update({ waterNo: waterNo }, { where: { id: id } });
-				await rooms.update({ waterMeterNo: waterMeterNo }, { where: { id: id } });
-				await rooms.update({ roomType: roomType }, { where: { id: id } });
-				await rooms.update({ status: status }, { where: { id: id } });
+				await rooms.update({ zoneId: UpdateRoomDTO(data).zoneId }, { where: { id: id } });
+				await rooms.update({ waterZoneId: UpdateRoomDTO(data).waterZoneId }, { where: { id: id } });
+				await rooms.update({ buildingId: UpdateRoomDTO(data).buildingId }, { where: { id: id } });
+				await rooms.update({ roomNo: UpdateRoomDTO(data).roomNo }, { where: { id: id } });
+				await rooms.update({ electricityNo: UpdateRoomDTO(data).electricityNo }, { where: { id: id } });
+				await rooms.update(
+					{ electricityMeterNo: UpdateRoomDTO(data).electricityMeterNo },
+					{ where: { id: id } }
+				);
+				await rooms.update({ waterNo: UpdateRoomDTO(data).waterNo }, { where: { id: id } });
+				await rooms.update({ waterMeterNo: UpdateRoomDTO(data).waterMeterNo }, { where: { id: id } });
+				await rooms.update({ roomType: UpdateRoomDTO(data).roomType }, { where: { id: id } });
+				await rooms.update({ status: UpdateRoomDTO(data).status }, { where: { id: id } });
 				return Response(res, SUCCESS_STATUS, NO_CONTENT_CODE);
 			} else {
 				return Response(res, INVALID_REFRESH_TOKEN, UNAUTHORIZED_CODE);
@@ -178,7 +152,110 @@ const updateRoom = async (req, res) => {
 	}
 };
 
+const createZone = async (req, res) => {
+	let data = await req.body;
+	try {
+		const zone = await zones.findOne({
+			where: {
+				name: data.zoneName,
+			},
+		});
+		const waterZone = await waterZones.findOne({
+			where: {
+				name: data.waterZoneName,
+			},
+		});
+		const building = await buildings.findOne({
+			where: {
+				name: data.buildingName,
+			},
+		});
+		// not create
+		if (zone && waterZone && building) {
+			return Response(res, SUCCESS_STATUS, OK_CODE, zone);
+		}
+		// create a new zone and water zone
+		else if (!!zone && !!waterZone && building) {
+			return Response(res, SUCCESS_STATUS, OK_CODE, 'Eznos');
+		}
+		// create a new zone and building
+		else if (!!zone && waterZone && !!building) {
+			return Response(res, SUCCESS_STATUS, OK_CODE, 'Eznos');
+		}
+		// create a new zone
+		else if (!!zone && waterZone && building) {
+			return Response(res, SUCCESS_STATUS, OK_CODE, 'Eznos');
+		}
+		// create a new water zone and building
+		else if (zone && !!waterZone && !!building) {
+			return Response(res, SUCCESS_STATUS, OK_CODE, 'Eznos');
+		}
+		// create a new water zone
+		else if (zone && !!waterZone && building) {
+			return Response(res, SUCCESS_STATUS, OK_CODE, 'Eznos');
+		}
+		// create a new building
+		else if (zone && waterZone && !!building) {
+			return Response(res, SUCCESS_STATUS, OK_CODE, 'Eznos');
+		}
+		// create all
+		else if (!!zone && !!waterZone && !!building) {
+			await zones.create({
+				name: data.zoneName,
+			});
+			const zone = await zones.findOne({
+				where: {
+					name: data.zoneName,
+				},
+			});
+			if (zone) {
+				await waterZones.create({
+					name: data.waterZoneName,
+				});
+
+				await waterZones.update(
+					{
+						zoneId: zone.id,
+					},
+					{ where: { name: data.waterZoneName } }
+				);
+
+				const waterZone = await waterZones.findOne({
+					name: data.waterZoneName,
+				});
+
+				await buildings.create({
+					name: data.buildingName,
+					lat: data.lat,
+					lng: data.lng,
+				});
+
+				await buildings.update(
+					{
+						zoneId: zone.id,
+					},
+					{ where: { name: data.buildingName } }
+				);
+
+				await buildings.update(
+					{
+						waterZoneId: waterZone.id,
+					},
+					{
+						where: { name: data.buildingName },
+					}
+				);
+
+				return Response(res, SUCCESS_STATUS, OK_CODE, 'ez');
+			}
+		}
+	} catch (err) {
+		return HandlerError(res, err);
+	}
+};
+
 module.exports.Building = building;
 module.exports.CreateRoom = createRoom;
 module.exports.DeleteRoom = deleteRoom;
 module.exports.UpdateRoom = updateRoom;
+module.exports.CreateZone = createZone;
